Homogenize shebang
Before the change, ansicat, netcache and opnk assume that the
interpreter location is always /bin/python.  While some distributions
do provide python at this location, even as a python3 interpreter,
some others such as Debian do not provide a plain python interpreter
anymore, at least not by default.  There exist packages to provide the
version 2 or the version 3 of the interpreter at that location, but
the three possible configurations (v2, v3 or none) are left at the
discretion of the administrator, so in practice can't really be
predicted in advance.

This change applies the same shebang as offpunk.py and the cache
migration script.  Use of /usr/bin/env has also the side effect of
easing use of python3 interpreters installed at weird locations.

Signed-off-by: Étienne Mollier <emollier@debian.org>