Importing PIL only if really needed
This is a small optimisation.

Python-pil is only needed if chafa is used with a version < 1.10.
(before 1.10, chafa would decode each frame of an animated gif,
python-pil was used to extract the first frame only)

But, if present, python-pil was always loaded despite never being used
if chafa >= 1.10 or if not chafa.

This commit thus only import python-pil if chafa is present and chafa <
1.10, saving memory and startup time.