Part two of the previous commit, adds the --ids option to netcache.
So, this is the part that doesn't want to apply. In the hope it is the problem, I replaced a directed ' with a straight '.
If that doesn't help, it is probably easiest to just copy and paste the code.
(The symbol does appear a few other times in the codebase but maybe those never passed through git send-email.)

Signed-off-by: Bert Livens <bert@bertlivens.be>