fix "list index out of range" error in get_link() when selecting 0
------------------------------
Good morning,

my first contribution here, also, 1st things 1st: I love it, especially because I never really managed to get wwwoffle running for the past 10 years or so.

Anyway, when on a site, and picking a link number <1, yo get a list index out of range error. The only condition was only to check if the index was too high. I simply check if the selected link is in the permitted range.

--------------------------------