adding some interactivity to cert management
These changes would improve (I hope) the workflow a bit when a page requires a
certificate. If offpunk is "offline", not much changes, except that we add a
message informing the user that they can run "certs" to provide a cert.

But, if online, it would interactively look if there is a cert for the site if
there are any, lets the user choose one. Or, create one if there aren't any.

It then applies the certificate directly without the user having to write "go
<certificate>@site", and fetches the page.

It can be tested by typing "go tilde.green/~jmcs/test/cert.sh", and choosing
among the options presented

(or, any gemini page that requires a certificate, for that matter) (this one
should ask for a cert, or echo back the "name" of the chosen cert)